+++ Models/Patient.cs
@@ -2,7 +2,7 @@ namespace Models
 {
     public class Patient
     {
-        public int PID { get; set; }
+        public long PID { get; set; }
         public string? PNAME { get; set; }
         public string? PEMAIL { get; set; }
         public bool PFOREIGN { get; set; }
