+++ Services/IPatientService.cs
@@ -5,6 +5,6 @@ namespace Services
     public interface IPatientService
     {
         Task<List<Patient>> GetAllAsync();
-        Task<Patient> GetByIdAsync(int id);
+        Task<Patient> GetByIdAsync(long id);
     }
 }
